requirements.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.1.3
django-countries==6.1.3
django-taggit==1.3.0
django-widget-tweaks==1.4.8
Faker==4.15.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
Pygments==2.7.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
wcwidth==0.2.5
